+++ contrib/scripts/check-crates.sh
@@ -41,6 +41,7 @@ buildargs=(
     "-p nostr --all-features"                                     # All features
     "-p nostr-database"
     "-p nostr-lmdb"
+    "-p nostr-mls-memory-storage"
     "-p nostr-mls-storage"
     "-p nostr-indexeddb --target wasm32-unknown-unknown"
     "-p nostr-ndb"
@@ -59,6 +60,7 @@ buildargs=(
 
 skip_msrv=(
     "-p nostr-lmdb"                       # MSRV: 1.72.0
+    "-p nostr-mls-memory-storage"         # MSRV: 1.74.0
     "-p nostr-mls-storage"                # MSRV: 1.74.0
     "-p nostr-keyring"                    # MSRV: 1.75.0
     "-p nostr-keyring --features async"   # MSRV: 1.75.0